Minimun design fixes
